backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
redis[asyncio]>=5.0.0
celery>=5.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.40
llama-index>=0.10.0
openai>=1.10.0
pinecone-client>=3.0.0
py-algorand-sdk>=2.4.0
beaker-pyteal>=1.1.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.7
python-dotenv>=1.0.0
pymupdf>=1.23.0
scikit-learn>=1.3.0
numpy>=1.24.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
aiohttp>=3.9.0
PyNaCl>=1.5.0
